% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_mixtures.R
\name{coef_mixtures}
\alias{coef_mixtures}
\title{Estimation of multinomial mixture distributions parameters}
\usage{
coef_mixtures(Mp, Np)
}
\arguments{
\item{Mp}{a matrix of mixtures (a row matrix if a vector)}

\item{Np}{a matrix of bases (a row matrix if a vector)}
}
\value{
This function returns a list of maximum likelihood estimates for the cells of a mixture distribution:

P matrix of cell probabilities estimates for mixtures

PI matrix of mixture probabilities

GAM matrix of cell probabilities estimates for bases
}
\description{
This function performs maximum likelihood inference for multinomial mixture distributions.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\references{
Yantis, S., Meyer, D. E., and Smith, J. E. K. (1991). Analyses of multinomial mixture distributions: New tests for stochastic models of cognition and action. Psychological Bulletin 110, 350–374.
}
\keyword{package}

