% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance_mixture.R
\name{deviance_mixture}
\alias{deviance_mixture}
\title{Deviance of multinomial mixture distributions}
\usage{
deviance_mixture(x, M, N, s, n, nbmel)
}
\arguments{
\item{x}{value to which the deviance is to be evaluated}

\item{M}{a vector of mixtures (see coef_mixtures.R)}

\item{N}{a vector of bases (see coef_mixtures.R)}

\item{s}{number of bases}

\item{n}{number of cell probabilities}

\item{nbmel}{number of mixtures}
}
\value{
This function returns the value of the deviance for mixture distributions.
}
\description{
This function calculates the deviance of multinomial mixture distributions.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\references{
Yantis, S., Meyer, D. E., and Smith, J. E. K. (1991). Analyses of multinomial mixture distributions: New tests for stochastic models of cognition and action. Psychological Bulletin 110, 350–374.
}
\keyword{package}

