% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expval_table.R
\name{expval_table}
\alias{expval_table}
\title{Expected values in a contingency table}
\usage{
expval_table(M)
}
\arguments{
\item{M}{a matrix of observed probabilities}
}
\value{
A matrix of expected values.
}
\description{
This function calculates expected values for a rxc contingency table.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\keyword{package}

