% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_test.R
\name{gof_test}
\alias{gof_test}
\title{Goodness-of-fit test for contingency tables}
\usage{
gof_test(lambda, observes, theoriques)
}
\arguments{
\item{lambda}{parameter defining the statistic to be used: lambda = -0.5 is for the Freeman-Tuckey statistic, lambda = 0 for the G2 statistic, lambda = 2/3 for the Cressie-Read statistic and lambda = 1 for the classical Chi-square statistic}

\item{observes}{vector of observed probabilities}

\item{theoriques}{vector of theoretical/expected probabilities}
}
\value{
This function returns the value of the goodness-of-fit statistic.
}
\description{
This function carries out goodness-of-fit tests for contingency tables from the power-divergence family.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\keyword{package}

