% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data_gen.R
\name{group_data_gen}
\alias{group_data_gen}
\title{Group individual capture-recapture data in encounter histories along specific column(s)}
\usage{
group_data_gen(X, effX, s)
}
\arguments{
\item{X}{matrix of capture-recapture histories}

\item{effX}{vector with numbers of individuals with that particular capture-recapture history}

\item{s}{scalar or vector of columns along which the grouping should be done}
}
\value{
matrix with grouped capture-recapture histories and counts in the last column
}
\description{
This function pools together individuals with the same encounter capture-recapture history along specified directions given by columns.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\keyword{package}

