% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_test_22.R
\name{ind_test_22}
\alias{ind_test_22}
\title{Test of independence for 2x2 contingency tables}
\usage{
ind_test_22(M, threshold = 2, rounding = 3)
}
\arguments{
\item{M}{is a 2x2 contingency table}

\item{threshold}{is a threshold for low expected numbers; default is 2}

\item{rounding}{is the level of rounding for outputs; default is 3}
}
\value{
This function returns a vector with statistic of quadratic chi2 or inv chi2 corresponding to pvalue of Fisher test, p-value of quadratic chi2 test or Fisher test for low numbers, signed test and test performed (Chi-square, Fisher or None).
}
\description{
This function tests independence in 2x2 contingency tables
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>,Jean-Dominique Lebreton, Rémi Choquet, Roger Pradel
}
\keyword{package}

