% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit_gen.R
\name{inv_logit_gen}
\alias{inv_logit_gen}
\title{Inverse generalized logit link}
\usage{
inv_logit_gen(petitv)
}
\arguments{
\item{petitv}{vector of values to be transformed}
}
\value{
ev vector of transformed values
}
\description{
This function computes the inverse (or reciprocal) of the generalized logit link function.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\keyword{package}

