% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimarray.R
\name{multimarray}
\alias{multimarray}
\title{Multistate m-array}
\usage{
multimarray(X, freq)
}
\arguments{
\item{X}{a matrix of encounter histories over K occasions}

\item{freq}{is a vector with the number of individuals having the corresponding encounter history}
}
\value{
This function returns a matrix in which R the number of released individuals is in the first column, the number of individuals never recaptured (K-1) is in the last column and m the m-array (K-1 x K-1) with upper triangle filled only is in sandwich between these two vectors.
}
\description{
This function calculates the m-array for multistate capture-recapture data, the number of released and never seen again individuals.
}
\examples{
# read in the geese dataset using package RMark
library(RMark)
geese = system.file("extdata", "geese.inp", package = "R2ucare")
geese = convert.inp(geese)

# add spaces between columns
geese.hist = matrix(as.numeric(unlist(strsplit(geese$ch, ''))),nrow=nrow(geese),byrow=T)
geese.freq = geese$freq

# encounter histories and number of individuals with corresponding histories
X = geese.hist
freq = geese.freq

# load R2ucare package
library(R2ucare)

# build m-array
multimarray(X,freq)
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Rémi Choquet, Jean-Dominique Lebreton, Anne-Marie Reboulet, Roger Pradel
}
\keyword{package}

