% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_ct.R
\name{pooling_ct}
\alias{pooling_ct}
\title{Pooling algorithm (multisite goodness-of-fit tests)}
\usage{
pooling_ct(table)
}
\arguments{
\item{table}{is a rxc contingency table}
}
\value{
This function returns a matrix with the pooled contingency table.
}
\description{
This function pools rows and columns of a rxc contingency table according to Pradel et al. (2003).
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Jean-Dominique Lebreton, Rémi Choquet, Roger Pradel
}
\references{
Pradel R., Wintrebert C.M.A. and Gimenez O. (2003). A proposal for a goodness-of-fit test to the Arnason-Schwarz multisite capture-recapture model. Biometrics 59: 43-53.
}
\keyword{package}

