% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling_mixtures.R
\name{pooling_mixtures}
\alias{pooling_mixtures}
\title{Pooling algorithm (multisite goodness-of-fit tests)}
\usage{
pooling_mixtures(nk, nj, a, mixandbases)
}
\arguments{
\item{nk}{number of mixtures}

\item{nj}{number of bases}

\item{a}{number of sites/states}

\item{mixandbases}{matrix with mixtures and bases}
}
\value{
This function returns a matrix with the pooled table.
}
\description{
This function pools rows and columns of a rxc bases and mixture table according to Pradel et al. (2003).
It provides the components of TestM in the multisite goodness-of-fit tests.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Rémi Choquet, Jean-Dominique Lebreton, Anne-Marie Reboulet, Roger Pradel
}
\references{
Pradel R., Wintrebert C.M.A. and Gimenez O. (2003). A proposal for a goodness-of-fit test to the Arnason-Schwarz multisite capture-recapture model. Biometrics 59: 43-53.
}
\keyword{package}

