% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstitution.R
\name{reconstitution}
\alias{reconstitution}
\title{Reformat outputs of multinomial mixture distributions parameters}
\usage{
reconstitution(x, s, n, nbmel)
}
\arguments{
\item{x}{vector with cell probabilities estimates for mixtures and bases, along with mixture probilities}

\item{s}{number of bases}

\item{n}{number of cell probabilities}

\item{nbmel}{number of mixtures}
}
\value{
This function returns a list of maximum likelihood estimates for the cells of a mixture distribution with:

P matrix of cell probabilities estimates for mixtures

PI matrix of mixture probabilities

GAM matrix of cell probabilities estimates for bases
}
\description{
This function reformat the outputs of multinomial mixture distributions parameters.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Roger Pradel, Rémi Choquet
}
\keyword{package}

