% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Replicate and tile a matrix}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{matrix to be replicated}

\item{m}{row dimension of replication}

\item{n}{column dimension of replication}
}
\value{
A replicated matrix of X with dimensions nrow(X)*m x ncol(X)*n.
}
\description{
This function creates a large matrix consisting of an m-by-n tiling of copies of X.
The dimensions of the returned matrix are nrow(X)*m x ncol(X)*n.
This is the equivalent of the repmat MATLAB function.
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>
}
\keyword{package}

