% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2ct.R
\name{test2ct}
\alias{test2ct}
\title{Test2.CT}
\usage{
test2ct(m, verbose = TRUE, rounding = 3)
}
\arguments{
\item{m}{is a triangular m-array (in a K-1 by K-1 array)}

\item{verbose}{controls the level of the details in the outputs; default is TRUE for all details}

\item{rounding}{is the level of rounding for outputs; default is 3}
}
\value{
This function returns a list with first component the overall test and second component a data.frame with 5 columns for components i (2:K-2) (in rows) of test2.Cti: component, degree of freedom, statistic of the test, p-value, test performed.
}
\description{
This function performs Test2.CT
}
\examples{
# read in Dipper dataset
library(RMark)
dipper = convert.inp('ed.inp',group.df=data.frame(sex=c('Male','Female')))

# add spaces between columns
dip.hist = matrix(as.numeric(unlist(strsplit(dipper$ch, ''))),nrow=nrow(dipper),byrow=T)
dip.freq = dipper$freq
dip.group = dipper$sex

# split the dataset in males/females
mask = (dip.group == 'Female')
dip.fem.hist = dip.hist[mask,]
dip.fem.freq = dip.freq[mask]
mask = (dip.group == 'Male')
dip.mal.hist = dip.hist[mask,]
dip.mal.freq = dip.freq[mask]

# for females
X = dip.fem.hist
freq = dip.fem.freq
res.fem = marray(X,freq)
m = res.fem$m[,,] # marray returns an array
res.females = test2ct(m)
res.females
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Rémi Choquet, Jean-Dominique Lebreton, Anne-Marie Reboulet, Roger Pradel
}
\keyword{package}

