% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test3Gwbwa.R
\name{test3Gwbwa}
\alias{test3Gwbwa}
\title{Test3G.WBWA}
\usage{
test3Gwbwa(X, freq)
}
\arguments{
\item{X}{is a matrix of encounter histories with K occasions}

\item{freq}{is a vector of the number of individuals with the corresponding encounter history}
}
\value{
This function returns a data.frame with occasion, site, the value of the test statistic, degree of freedom, p-value and test performed (chi-square, Fisher or none).
}
\description{
This function performs Test3G.WBWA
}
\examples{
# read in Geese dataset
library(RMark)
geese = system.file("extdata", "geese.inp", package = "R2ucare")
geese = convert.inp(geese)

geese.hist = matrix(as.numeric(unlist(strsplit(geese$ch, ''))),nrow=nrow(geese),byrow=T)
geese.freq = geese$freq

# encounter histories and number of individuals with corresponding histories
X = geese.hist
freq = geese.freq

# load R2ucare package
library(R2ucare)

# perform Test.3GWBWA
test3Gwbwa(X,freq)
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Rémi Choquet, Roger Pradel
}
\keyword{package}

