% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool2K.R
\name{pool2K}
\alias{pool2K}
\title{Pooling algorithm}
\usage{
pool2K(M, low = 2)
}
\arguments{
\item{M}{is a 2 by K contingency table (or a K by 2 table)}

\item{low}{is a threshold for low expected numbers; default is 2 (if this argument is big enough, the table is pooled down to 2 x 2; if this argument is 0, the table is not pooled)}
}
\value{
This function returns a matrix with the pooled contingency table.
}
\description{
This function pools columns of a 2xK contingency table (if needed, ie if low numbers present)
}
\author{
Olivier Gimenez <olivier.gimenez@cefe.cnrs.fr>, Jean-Dominique Lebreton, Rémi Choquet, Roger Pradel
}
\keyword{package}

